examples/news_factor/requirements.txt
# 核心量化框架 (使用本地开发版本)
# qlib>=0.9.0  # 注释掉PyPI版本
-e ../../  # 使用本地qlib开发版本

# 数据处理
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
TA-Lib>=0.4.29

# 加密货币交易
ccxt>=4.0.0

# 新闻数据采集
feedparser>=6.0.8
requests>=2.25.1
beautifulsoup4>=4.9.3
scrapy>=2.5.0

# 情感分析
transformers>=4.20.0
torch>=1.11.0
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.7

# 机器学习
scikit-learn>=1.0.0
xgboost>=1.6.0
lightgbm>=3.3.0

# 数据存储
pymongo>=4.0.0
redis>=4.0.0
sqlalchemy>=1.4.0

# 配置管理
pyyaml>=6.0
python-dotenv>=0.19.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 日志和监控
loguru>=0.6.0

# 并发处理
asyncio>=3.4.3
aiohttp>=3.8.0

# 时间处理
python-dateutil>=2.8.2
pytz>=2021.3

# 数据验证
pydantic>=1.8.0

# 测试
pytest>=6.2.0
pytest-asyncio>=0.18.0

# API接口
fastapi>=0.75.0
uvicorn>=0.17.0

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# 开发工具
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.950 